+++ lib/import/resources.rb
@@ -27,10 +27,11 @@ module Import
       log "Importing resources from EOSC Registry..."
 
       begin
+        token = Importers::Token.new(faraday: @faraday).receive_token
         response =
-          Importers::Request.new(@eosc_registry_base_url, "resource/rich", faraday: @faraday, token: @token).call
-      rescue Errno::ECONNREFUSED
-        abort("import exited with errors - could not connect to #{@eosc_registry_base_url}")
+          Importers::Request.new(@eosc_registry_base_url, "resource/rich", faraday: @faraday, token: token).call
+      rescue Errno::ECONNREFUSED, Importers::Token::RequestError => e
+        abort("import exited with errors - could not connect to #{@eosc_registry_base_url} \n #{e.message}")
       end
 
       updated = 0